pyproject.toml
[tool.poetry]
name = "reactionmodel"
version = "0.1.0"
description = "A package for defining a set of reactions which convert reactants to products and easily simulating them deterministically or stochastically."
authors = ["Thayer Anderson <tanderson11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25"
simpleeval = "^0.9.13"
pandas = "^2.1.4"
scipy = "^1.11.4"
pyyaml = "^6.0.1"

[tool.poetry.group.extras.dependencies]
numba = "^0.58.1"

[tool.poetry.group.dev.dependencies]
ipython = "^8.17.2"
jupyter = "^1.0.0"
matplotlib = "^3.8.3"
pylint = "^3.1.0"
ruff = "^0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
